Okay, this is a huge commit!!! I really need to get in to the habbit of making small changes and then committing them!
I have finally implemented Redux!!!

I have completely restructured the whole project, the redux elements are 
now distributed amongst the folders: actions & reducers, the store is 
declared in store.js. The components have been devided in to smart 
components in the folder: pages and dumb components in the folder: 
components. I have moved the Node server .js files in to the folder: 
server. There is also a folder named: utils that will be the home of 
certain functions. At the moment it has the serverRequests.js function 
in.

The previous and next buttons in the single image component now work 
although I am going to have to change them a bit. At the mo they 
re-render the whole page when you click them, the same with when you 
click on the individual images in the gallery view. I need to change the 
<a href=> links to use React-Routers <link to=> method. I think that 
should fix it although i will have to change the whole parameter passing 
functionality around a bit, It'll probably be better of afterwards 
anyway, Hopefully that will be my next commit.

I have also installed the package redux-thunk so I can use asyncronous 
functions in the redux actions but I didn't use it in the end so I may 
remove it later.