Fixed the issue where files weren't getting displayed because of the characters in their filenames.
Turnes out, it was just the filenames with '&' in and the problem was 
the URLSearchParams.get method would stop when it gets to an '&' symbol. 
I fixed it by just relying on the index. I have stripped the filename 
parameter from the url. The filename is still displayed though just to 
make the url look cohearent.