The website deploys now, i've had to make some changes to all the URLs do that they are relative.
I have also made my serverRequest function slightly more dynamic, you 
don't need to give it the server's url now just the request type 
although it should still work if want to give it the full URL.

The site does not work in dev mode now, the links are all wrong! I just 
need to try and get the site to work the same in dev mode as it does 
when deployed.

Hopefully that will be the next commit.