Got the basic add to basket and remove from basket working.
Lots of stuff not implemented yet like: 

In gallery view the "add to basket" need to change to "remove" if it is 
already in the basket.

There is nothing stopping you from adding the same file multiple times. 
(possibley this oroblem will be fixed with the above problem)

No download button implemented yet.

When you view an individual file in the basket, the next and prev 
buttons still navigate through the whole list of jpegs not just the ones 
in the basket.

There's probably more stuff!!!