The add button now changes to remove which fixed the problem of being able to add the same image multiple times.
The basket button in the headdar now reports the number of items in the 
basket.

I have tweeked the CSS a little to float the basket button on the right

I have also added an origin perameter to the url indicating to the 
single view weather or not to display images from the gallery or the 
basket.