I was mistaken in the last commit, i think this one fixes it.
The reason I was mistaken is that I accidently used the online version 
when I was testing the full size feature. It looks like it's working 
fine on the online version now and I havent built the full size feature 
in to it yet. I might end up scrapping this commit if that is the case 
but here it is anyway.