Cleared up some bugs IE:
next an prev buttons in single view will change to end and start at the 
beginning and end of the basket or gallery. 

The remove button in single view is now also a link to the previous 
image in the basket or to an empty basket if it's the last image in the 
basket.

tweeked the CSS so the remove link still looks like a button. Still 
needs a lot of work styling.

I am finishing for the day now but I cant help feeling like I might want 
to optimise the code I've added today, there's a lot of it and even a 
small ammount of repetition. I'll leave that for another day though.

as for now everything is pretty consistent, i'm going to merge with 
Master which will finalise this branch.

The next project will be to add the download button, I'll start a new 
branch for that.