Right, this is another large commit, perhaps a bit to large! Apologies.
The download button works!!!

It was a slow road though and I made a lot of changes along the way.

I am no longer using node package adm-zip, I have swapped that for
archiver.

The galleryServer zipJpegs route no longer saves a file. It pipes a
write stream back to the browser which will initiate a save-file-as
prompt. It's better this way, there is no temp file on the server and
(although not tested) I'd imaging multiple users could DL a zip at the
same time without problem.

Finally, there is no longer a fetch request for zipJpegs! It's obvious
really but this was why I spent so long trying to work it out. the
zipJpegs function in the serverRequests file now just returns the url
to the <a href link attached to the download button.