I have added a download all button Too!
The gallery button now switches to a DL-All button when in gallery view.

The button picker function in the header is now dealing with the basket
button and the gallery button so I used a switch case statement.

I also now have to pass getJpegs function in to the gallery component so
it can in turn pass it to the header button.