I've rewritten the single image view so that it's not dependant on urlParams. It only needs the file name now and the origin is embedded in the URL which is accessed using the url.split('/') method. I have also had to pass the setOrigin() method down through the parameters so it can be part of the URL parameters.
The next and previous functions now just change the index of the file
in the array of jpegsSource. Before they were a React-Router link to
the new file with all the parameters in the searchParams part of the
URL.

I am now using React hooks to trigger re-rendering when you change the
file index so it changes the image.

Next thing to do is to implement React-Springs useTransition method and
see if I can get the image transitions animated.

Stay Tuned!!!