Changed the data flow to make it more logical.
It's all about having the id field. Before the ID was set on the server
and passed to the browser to be stored. This meant that the ADD_JPEG
function method just needed to add the whole response from the server to
it's jpegs. Consequently the add basket logic had to change a fair bit
so that it was adding an ID value to each file that was added to the
basket. This introduced a bunch of other issues like keeping track of
an ID counter and keeping the numbers sequential after items were
removed.

I have changed it now so that the server only returns a list of jpegs
(like it did in the beginning), they are added one at a time to the
state and given an ID at the time of being added. All other processes
now deal with jpeg items that include the file name and the id.

The whole thing makes much more sense this way