Bug Fixes:
ADD_JPEGS now can handle an array of jpegItems. it's quicker create the
data object first and pass it to redux rather than doing it one file at
a time.

The getJpegs serverRequest now wraps the array of files from the server
in to an array of objects that contain each file coupled with an id.

Add All To Basket button now works

The image slider is now capable of handling filenames with "(" and ")"
in.

Fixed a little bug in the thumbnails component where it was using the i
variable.