I have had to expand all the function buttons out so there is now a function bar with "download", "add all" and the new "empty basket" button (which does not work yet. I'm thinking about adding another Redux action for it).
The original navbar buttons new have a nice hover effect and they also
reflect which page you are on.

I have to finish for the day but there is a load of stuff that needs
deleting in header.js. All the old buttons are still there and I have 
dramatically changed the button picker function.
 I will address this tomorrow.