This should probably be 2 separate commits, my apologies!
Part 1: --
Added new actions to the basket reducer:

EMPTY_BASKET - Does what it says!

ADD_ALL - Receives an array and replaces the whole basket with the
array. I introduced this action because it was so slow iterating
through the array and adding the jpegs one by one. This is super quick
now!

Part 2: --
Tidied up the header. I have put all the button logic in the button
picker function. It doesn't seem to slow my app down and it makes the
whole thing more intuitive and readable.

altered the way the function buttons appear slightly. after much
deliberation I decided it looks better to have them visible all the
time and just ghosted out when their not applicable.

I have had to pass more of the props through to more components to get
the buttons visible in components that they were not visible in before.

Next I am going to try and move the header back to the app component so
I will probably be able to remove a lot of the props routing.