The Header now works beautifully!!!
I have had to convert the header component from a stateless function
component to a React.Component. There is now a state in App.js that is
tracking the selectedPage so the header can update it's buttons
accordingly.

It might be worth mentioning that the buttons no longer rely on 'origin'
search parameter that was being used to track the entry point for the
single image view (I.e. the basket or the gallery).

Also the only time the URL is used is to initially set the state when
the page is first loaded.