Have found a bug!
The remove item button in single image view randomly crashes!!!

To be more specific it randomly fails to

I have added loads of debug console logs to singleImage.js but I'm
reluctant to spend a lot of time trying to fix it.

Ideas for fix (need to experiment):

1 - make next and prev in to hooks and have them update with useEffect.

2 - Type Check all the props with propTypes (This should be my first
experiment)

3 - It could be to do with the URL file name not matching the filename
in props.getBasket although I checked this and it doesn't seem like it

Going to leave as is for now!