Still not isolated the bug!
Added more debug console.logs

The problem is that sometimes for reasons unknown jpegsArray in
singleImage is not directed to the correct array (basket or gallery).

I've tried to use useEffect hooks to trigger the appropriate code but to
no avail! The attempt is commented out for now.

I have implemented prop checks in singleImage with propTypes.

Going to crack on with the styling for now.