The bug!!!
I have made the bug better by introducing a 10ms delay before the
removeJpeg action is dispatched. It still crashes but it's much less
common with the 10ms delay, I'd even say it's rare!

Strangely the crash is more common if you delay for 1 ms or 1 second!
10ms seems to be a sweet spot!

I'm going to deploy it and see how it behaves online.