The Re-sequence Jpegs function is now in a new file called utils.js in the utils folder. It is now imported to the gallery and basket components and applied on the fly.
When I was experimenting with the re-sequence Jpegs routine in the
serverRequests file and applying it to the list of jpegs returned from
the server. I discovered a bug where I had to refresh the page to get
the re-sequencing to happen! I managed to fix this by making the nested
function that adds id to the jpegs array async and making it await the
responseData.map() method.

Although I am not using the re-sequence Jpegs function here anymore I
have left this in as It seems like it should be awaiting that whole
process. I wondered if this would fix the async bug that I have
encountered but it did not. might be worth checking for more nested
functions like this that could be awaited though.