Trying to get the browser to wait for the Image()
Iv've had a go trying to get the browser to wait for Image() to return
the width and height data but unfortunately to no avail. There are two
commented out .then() statements in the getJpegs method that almost work
but not quite. I'm going to have to move on for time constraint reason.