Re-sequence dynamically on viewport change
I have added a Media Query event listener to listen for when the
viewport width changes to below or above 800px. When it does it calls
resequenceJpegs().

I have had contain the jpegs in useState so that i can re-render the DOM
when their sequence changes.

I've also tidied up some of the comments in resequenceJpegs()

This commit doesn't quite work. You have to click on off and on the
gallery button to get it to display the images. I'm going to fix that in
the next commit.