Working! Dynamic viewport size change & re-sequence
I have managed to get serverRequest.js getJpegs to properly wait while
it looks up the resolution of each image. I had to wrap each callback
from the map() iterator in a promise as well as the whole iterator. I'm
going to leave all my redundant code there for a while until I'm
satisfied I will not be needing it in the future so I will have to
clean it up at some point.

I have also introduced an "isFetching" variable in the main state for
the jpegs object and have incorporated it in to conditional returns when
rendering the gallery. I'm hoping this could also be the solution to
fixing the "async bug" I'm having when you remove the image from the
basket in single image basket view. I have yet to implement this though.

There is now a new fileReducer Action "toggleIsFetching(boolean)" and a
new state subscription "isFetchingJpegs".