Major overhaul of single image view
Single image view is now vastly different!

I have removed a couple of containers that were superfluous and created
a new formatting context for the whole image section.

I have arranged the next and prev buttons so that they take up 1/3 of
the image element on either side, they now have left and right chevrons
as icons.

I have added an exit button on the top left and the add/remove button
now sits on the top right all with tasty SVG icons.

I have removed the header spacer element and replaced it with margins.
I think this still needs a bit of tweaking!