Async Bug FIXED!!!!!!
I have no idea how but at some point over the last few commits the
random crash from removing the image from the basked in single image
view (I've been referring to it as the async bug) has been resolved!!!
My best guess as to how is that I have simplified the button logic for
the remove button, specifically I placed the onClick event that
dispatched the removeBasked action in to the link. Before there was a
button embedded in the link.

I have removed the delay from the removeBasket Redux action!