Attempting to make thumbnails fade out!
I have added a fade out CSS animation and I am attempting to make
thumbnails in the basket fade out when you remove them from the basket.
I need to create a hook for the fade variable but being as the thumbnail
component is already distributed in a loop it complains.

I have commented out the useStat line until I have less important things
to do. I'm going to work on my welcome page for now.