Thumbnails Fade Out When Removed from Basket!
I have added a new field to the jpeg object. When getJpegs imports the
jpegs array it now also adds the field "toBeDeleted" = false. I am using
this to trigger changing the 'fade' variable in the thumbnails
component that sets the className to either 'fade-in' or 'fade-out'.

I have also had to add a new basket reducer "TO_BE_REMOVED" and the
action to remove jpegs now dispatches that first, it then waits for a
few milliseconds for the animation to play out before it actually
removes the item from the basket.

I also deleted some redundant code from serverRequests.