Finished About css
Tidied up CSS and html classnames

Responsive css is now much better

I discovered what was causing the overspill in mobile view, it was some
padding on the home-container. I fixed it by subtracting it from the
element width.

I'm hoping I won't have to mess with css again for the about page.