Added re-sequencing for delete
I have reshuffled the structure a bit. The resequenceJpegs() function is
now called by the reducer when the images are added to the basket
instead of when the page is re-rendered.

I have also added a new resequenceDelete() function that insures there
minimum reshuffling of the thumbnails when in gallery view.

I have renamed utils.js to sequencer.js because all it had in it was
re-sequencing code.