Added keyboard navigation to single viewer
I have added a simple event listener to single image viewer that will
react to arrow left and right for prev and next images and escape to
exit.

It occurs to me tho that once I started using it is felt very unnatural
to not be able to use the keys after exiting! I think it would be a good
idea to implement keyboard navigation throughout the whole web app.